+++ crates/nu-command/src/commands/termsize.rs
@@ -35,13 +35,23 @@ impl WholeStreamCommand for TermSize {
         match size {
             Some((w, h)) => {
                 if wide && !tall {
-                    Ok(ActionStream::one(UntaggedValue::int(w).into_value(tag)))
+                    Ok(ActionStream::one(
+                        UntaggedValue::int(w as i64).into_value(tag),
+                    ))
                 } else if !wide && tall {
-                    Ok(ActionStream::one(UntaggedValue::int(h).into_value(tag)))
+                    Ok(ActionStream::one(
+                        UntaggedValue::int(h as i64).into_value(tag),
+                    ))
                 } else {
                     let mut indexmap = IndexMap::with_capacity(2);
-                    indexmap.insert("width".to_string(), UntaggedValue::int(w).into_value(&tag));
-                    indexmap.insert("height".to_string(), UntaggedValue::int(h).into_value(&tag));
+                    indexmap.insert(
+                        "width".to_string(),
+                        UntaggedValue::int(w as i64).into_value(&tag),
+                    );
+                    indexmap.insert(
+                        "height".to_string(),
+                        UntaggedValue::int(h as i64).into_value(&tag),
+                    );
                     let value = UntaggedValue::Row(Dictionary::from(indexmap)).into_value(&tag);
                     Ok(ActionStream::one(value))
                 }
